# perf: delay guess encoded