--- a/.eslintrc
+++ b/.eslintrc
@@ -31,7 +31,7 @@
     },
     "extends": "eslint:recommended",
     "parserOptions": {
-        "ecmaVersion": 6,
+        "ecmaVersion": 2019,
         "sourceType": "module"
     }
 }