--- a/.eslintrc
+++ b/.eslintrc
@@ -31,7 +31,7 @@
     },
     "extends": "eslint:recommended",
     "parserOptions": {
-        "ecmaVersion": 2019,
+        "ecmaVersion": 2020,
         "sourceType": "module"
     }
 }