*This PR includes changes from #215*

**Potentially breaking change**
I don't think `MagicString::indentStr` is public because it is not written on docs nor type definitions.
But if someone is relying on this, it would break.

---

This PR delays executing `guessIndent(originalContent)`.
This takes slightly long when the content is large.

**before**
```
node v16.15.0

construct x 5,154 ops/sec ±0.96% (95 runs sampled)
append x 7,987,412 ops/sec ±3.90% (62 runs sampled)
indent x 4,177 ops/sec ±0.57% (91 runs sampled)
generateMap (no edit) x 1,894 ops/sec ±0.95% (89 runs sampled)
generateMap (edit) x 1,843 ops/sec ±0.40% (93 runs sampled)
generateDecodedMap (no edit) x 1,954 ops/sec ±1.00% (91 runs sampled)
generateDecodedMap (edit) x 1,959 ops/sec ±0.59% (92 runs sampled)
```
**after**
```
node v16.15.0

construct x 80,654 ops/sec ±0.53% (93 runs sampled)
append x 7,852,165 ops/sec ±5.47% (58 runs sampled)
indent x 3,785 ops/sec ±0.53% (92 runs sampled)
generateMap (no edit) x 1,983 ops/sec ±0.50% (96 runs sampled)
generateMap (edit) x 1,979 ops/sec ±0.53% (94 runs sampled)
generateDecodedMap (no edit) x 2,185 ops/sec ±0.41% (94 runs sampled)
generateDecodedMap (edit) x 2,176 ops/sec ±0.45% (94 runs sampled)
```