@antfu 2022-06-04
Can you make this PR orthogonal to #215 so we could have it first? For #215 I am discussing with Rich to have upstream improvements directly.