@sapphi-red 2022-06-05
I made this orthogonal to it. 👍